requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0
